chore: codeorg: move pcie related code in pkg/pcie
We want to introduce a cpuenumerator object which will
be responsible to map CPU devices to DRA devices, removing
this code from the CPUDriver (and from dra_hooks.go).

To simplify the dependency graph, we move the pcie processing
code in its own package.
Trivial code movement and import fixes.
