+++ pkg/pcie/pciescan_amd64_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package device
+package pcie
 
 import (
 	"io/fs"
