+++ pkg/pcie/testdata_ryzen9_5950x.go
@@ -4,7 +4,7 @@
 // regenerate from the project root with:
 // go run tools/gen-pcie-testdata/main.go --func makeRyzen95950XFixture --pkg device --build-tag amd64
 
-package device
+package pcie
 
 import (
 	"testing/fstest"
